+++ cmd/osm-controller/osm-controller.go
@@ -215,7 +215,7 @@ func main() {
 		cfg,
 		endpointsProviders...)
 
-	proxyRegistry := registry.NewProxyRegistry()
+	proxyRegistry := registry.NewProxyRegistry(&registry.KubeProxyServiceMapper{KubeController: kubernetesClient})
 	proxyRegistry.ReleaseCertificateHandler(certManager)
 
 	adsCert, err := certManager.IssueCertificate(xdsServerCertificateCommonName, constants.XDSCertificateValidityPeriod)
